+++ pom.xml
@@ -951,8 +951,8 @@
                             <arg>-feature</arg>
                             <arg>-explaintypes</arg>
                             <arg>-Yno-adapted-args</arg>
-                            <arg>-Ywarn-unused:imports</arg>
-                            <arg>-Xlint:missing-interpolator</arg>
+			    <!--<arg>-Ywarn-unused:imports</arg>
+			    <arg>-Xlint:missing-interpolator</arg>-->
                             <arg>-Xfatal-warnings</arg>
                         </args>
                         <jvmArgs>
