+++ sql-plugin/src/main/321+-db/scala/com/nvidia/spark/rapids/shims/DeltaLakeUtils.scala
