+++ sql-plugin/src/main/scala/org/apache/spark/sql/rapids/complexTypeExtractors.scala
@@ -26,6 +26,7 @@ import com.nvidia.spark.rapids.shims.ShimUnaryExpression
 import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.catalyst.analysis.TypeCheckResult
 import org.apache.spark.sql.catalyst.expressions.{ExpectsInputTypes, Expression, ExtractValue, GetArrayStructFields, ImplicitCastInputTypes, NullIntolerant}
+import org.apache.spark.sql.catalyst.trees.TreePattern._
 import org.apache.spark.sql.catalyst.util.{quoteIdentifier, TypeUtils}
 import org.apache.spark.sql.rapids.shims.RapidsErrorUtils
 import org.apache.spark.sql.types.{AbstractDataType, AnyDataType, ArrayType, BooleanType, DataType, IntegralType, MapType, StructField, StructType}
@@ -43,6 +44,8 @@ case class GpuGetStructField(child: Expression, ordinal: Int, name: Option[Strin
     val fieldName = if (resolved) childSchema(ordinal).name else s"_$ordinal"
     s"$child.${name.getOrElse(fieldName)}"
   }
+
+  override def nodePatternsInternal(): Seq[TreePattern] = Seq()
 
   override def sql: String =
     child.sql + s".${quoteIdentifier(name.getOrElse(childSchema(ordinal).name))}"
@@ -95,6 +98,8 @@ case class GpuGetArrayItem(child: Expression, ordinal: Expression, failOnError: 
   override def dataType: DataType = child.dataType.asInstanceOf[ArrayType].elementType
 
   override def hasSideEffects: Boolean = super.hasSideEffects || failOnError
+
+  override def nodePatternsInternal(): Seq[TreePattern] = Seq()
 
   override def doColumnar(lhs: GpuColumnVector, rhs: GpuColumnVector): ColumnVector = {
     val (array, indices) = (lhs.getBase, rhs.getBase)
@@ -332,6 +337,8 @@ case class GpuGetArrayStructFields(
   override def dataType: DataType = ArrayType(field.dataType, containsNull)
   override def toString: String = s"$child.${field.name}"
   override def sql: String = s"${child.sql}.${quoteIdentifier(field.name)}"
+  
+  override def nodePatternsInternal(): Seq[TreePattern] = Seq()
 
   override protected def doColumnar(input: GpuColumnVector): ColumnVector = {
     val base = input.getBase
